Move array death test into a separate serially executed test
Death tests are known to have issues when threads are involved. A better
explanation is provided as part of google-test documentation at the
below link.

https://github.com/google/googletest/blob/master/docs/advanced.md#death-tests-and-threads
(cherry picked from commit 6633e108fc4b2f201d032f092e1d5845951c2ad5)
