+++ test/CMakeLists.txt
@@ -238,6 +238,7 @@ make_test(SRC anisotropic_diffusion.cpp)
 make_test(SRC approx1.cpp)
 make_test(SRC approx2.cpp)
 make_test(SRC array.cpp CXX11)
+make_test(SRC array_death_tests.cpp CXX11 SERIAL)
 make_test(SRC arrayio.cpp)
 make_test(SRC assign.cpp CXX11)
 make_test(SRC backend.cpp CXX11)
